Changed from home activity to home fragment
